refactor: WinningLotto -> WinLotto 클래스명 수정 및 초기값 타입 수정
